+++ src/AppserverIo/Appserver/Application/Application.php
@@ -26,6 +26,7 @@ use AppserverIo\Storage\GenericStackable;
 use AppserverIo\Storage\StorageInterface;
 use AppserverIo\Appserver\Core\Utilities\DirectoryKeys;
 use AppserverIo\Appserver\Core\Traits\ThreadedContextTrait;
+use AppserverIo\Appserver\Core\Interfaces\ContainerInterface;
 use AppserverIo\Appserver\Core\Interfaces\ClassLoaderInterface;
 use AppserverIo\Appserver\Core\Api\Node\ContextNode;
 use AppserverIo\Appserver\Core\Api\Node\ManagerNodeInterface;
@@ -204,7 +205,7 @@ class Application extends \Thread implements ApplicationInterface, DirectoryAwar
      */
     public function getAppBase()
     {
-        return $this->getNamingDirectory()->search('php:env/appBase');
+        return $this->getNamingDirectory()->search(sprintf('php:env/%s/appBase', $this->getName()));
     }
 
     /**
@@ -430,11 +431,12 @@ class Application extends \Thread implements ApplicationInterface, DirectoryAwar
      * Prepares the application with the specific data found in the
      * passed context node.
      *
-     * @param \AppserverIo\Appserver\Core\Api\Node\ContextNode $context The application configuration
+     * @param \AppserverIo\Appserver\Core\Interfaces\ContainerInterface $container The container instance bind the application to
+     * @param \AppserverIo\Appserver\Core\Api\Node\ContextNode          $context   The application configuration
      *
      * @return void
      */
-    public function prepare(ContextNode $context)
+    public function prepare(ContainerInterface $container, ContextNode $context)
     {
 
         // load application name + naming directory
@@ -445,17 +447,19 @@ class Application extends \Thread implements ApplicationInterface, DirectoryAwar
         $namingDirectory->createSubdirectory(sprintf('php:global/%s', $applicationName));
 
         // create the applications 'env' + 'env/persistence' directory the beans + persistence units will be bound to
+        $namingDirectory->createSubdirectory(sprintf('php:env/%s', $applicationName));
         $namingDirectory->createSubdirectory(sprintf('php:global/%s/env', $this->getName()));
         $namingDirectory->createSubdirectory(sprintf('php:global/%s/env/persistence', $this->getName()));
 
+        $namingDirectory->bind(sprintf('php:env/%s/appBase', $this->getName()), $container->getAppBase());
+
         // prepare the application specific directories
-        $webappPath = sprintf('%s/%s', $namingDirectory->search('php:env/appBase'), $applicationName);
-        $tmpDirectory = sprintf('%s/%s', $namingDirectory->search('php:env/tmpDirectory'), $applicationName);
+        $webappPath = sprintf('%s/%s', $this->getAppBase(), $applicationName);
+        $tmpDirectory = sprintf('%s/%s', $container->getTmpDir(), $applicationName);
         $cacheDirectory = sprintf('%s/%s', $tmpDirectory, ltrim($context->getParam(DirectoryKeys::CACHE), '/'));
         $sessionDirectory = sprintf('%s/%s', $tmpDirectory, ltrim($context->getParam(DirectoryKeys::SESSION), '/'));
 
         // prepare the application specific environment variables
-        $namingDirectory->createSubdirectory(sprintf('php:env/%s', $applicationName));
         $namingDirectory->bind(sprintf('php:env/%s/webappPath', $applicationName), $webappPath);
         $namingDirectory->bind(sprintf('php:env/%s/tmpDirectory', $applicationName), $tmpDirectory);
         $namingDirectory->bind(sprintf('php:env/%s/cacheDirectory', $applicationName), $cacheDirectory);